Modified the landing CAEQ and added a new screen for public directory
